move self.engine with property
