+++ pgdrive/component/map/base_map.py
@@ -78,7 +78,6 @@ class BaseMap(BaseObject):
         self.blocks = []
 
         # Generate map and insert blocks
-        self.engine = get_engine()
         self._generate()
         assert self.blocks, "The generate methods does not fill blocks!"
