+++ pgdrive/component/vehicle/_deprecated_traffic_vehicle_type.py
