+++ pgdrive/utils/random.py
@@ -7,6 +7,7 @@ import os
 import struct
 import uuid
 from typing import Optional
+
 import numpy as np
 
 
@@ -66,7 +67,7 @@ def create_seed(a=None, max_bytes=8):
         a += hashlib.sha512(a).digest()
         a = _bigint_from_bytes(a[:max_bytes])
     elif isinstance(a, int):
-        a = a % 2**(8 * max_bytes)
+        a = a % 2 ** (8 * max_bytes)
     else:
         raise logging.error('Invalid type for seed: {} ({})'.format(type(a), a))
 
@@ -81,7 +82,7 @@ def _bigint_from_bytes(bytes):
     unpacked = struct.unpack("{}I".format(int_count), bytes)
     accum = 0
     for i, val in enumerate(unpacked):
-        accum += 2**(sizeof_int * 8 * i) * val
+        accum += 2 ** (sizeof_int * 8 * i) * val
     return accum
 
 
@@ -94,7 +95,7 @@ def _int_list_from_bigint(bigint):
 
     ints = []
     while bigint > 0:
-        bigint, mod = divmod(bigint, 2**32)
+        bigint, mod = divmod(bigint, 2 ** 32)
         ints.append(mod)
     return ints
 
@@ -117,6 +118,12 @@ class RandomEngine:
 
     def randint(self):
         return self.np_random.randint(0, self.MAX_RAND_INT)
+
+    @property
+    def engine(self):
+        """All object can access the engine with self.engine!"""
+        from pgdrive.utils.engine_utils import get_engine
+        return get_engine()
 
 
 def random_string(prefix=None):
